+++ src/databar/__init__.py
@@ -101,7 +101,7 @@ from .models import (
     Folder,
 )
 
-__version__ = "2.3.0"
+__version__ = "2.4.0"
 __all__ = [
     "DatabarClient",
     # exceptions